requirements.txt
PyMuPDF
fitz
openai
python-dotenv
requests
tqdm